Move layer configuration munging to route
