+++ app/templates/components/map-from-id.hbs
@@ -7,7 +7,7 @@
   mapLoaded=(action 'handleMapLoad')
   as |map|}}
 
-  {{#each builtLayers key='id' as |layer|}}
+  {{#each visibleLayers key='id' as |layer|}}
     {{map.layer layer=layer before='waterway-label'}}
   {{/each}}
